backend/requirements.txt
# --- Core Python / Flask Framework ---
Flask>=3.0.0
Flask-Cors>=4.0.0  # Package name uses hyphen, import uses underscore
python-dotenv>=1.0.1
gunicorn>=21.0.0  # For deployment later

# --- Data Validation / Models ---
pydantic>=2.5.0,<3.0.0  # Version compatible with Python 3.11/3.12

# --- LlamaIndex Core ---
# Specify a recent stable core version. Integrations will align.
llama-index-core>=0.10.50  # Or use llama-index>=0.10.50 if core causes issues

# --- LlamaIndex Integrations (let pip find compatible versions) ---
llama-index-llms-google-genai
llama-index-embeddings-huggingface  # For local build script & runtime query
llama-index-vector-stores-faiss
llama-index-graph-stores-neo4j

# --- Direct Dependencies for Integrations & Build Script ---
google-generativeai>=0.5.0  # For the LLM
sentence-transformers>=2.6.0  # For local embeddings
torch>=2.0.0  # Dependency for sentence-transformers
faiss-cpu>=1.7.0  # FAISS engine
neo4j>=5.10.0  # Neo4j Python driver
numpy>=1.24.0  # Often needed by faiss/embeddings

# --- Firebase / Google Cloud ---
firebase-admin>=6.3.0
google-cloud-firestore>=2.11.0

# --- Authentication ---
PyJWT>=2.8.0

# --- Optional Utilities ---
requests>=2.31.0  # Not strictly needed by backend itself
